# Optimizations: `tt.shared.join`, several `ttsim.interface_dag_elements.fail_if` functions