@JuergenWiemers 2025-08-13
Two more thoughts:
- I should add every existing GETTSIM target leaf to the `tt_targets` and override none of the intermediate nodes in the mapper in my benchmark script. This way the script should be able to catch badly scaling code in any GETTSIM policy function or in TTSIM.
- In order to detect possible runtime/scaling regressions in GETTSIM/TTSIM that may be accidentally introduced in a future PR, it might even be useful to perform this type of benchmark (semi-)automatically. For example, the maintainers of the Julia language have set up a Github bot they call "nanosoldier". If called in a PR comment with the keyword `@nanosoldier`, it runs a large benchmark suite on the most popular packages in the Julia ecosystem, once for the PR and once for the main branch and compares results. Here is a recent [example](https://github.com/JuliaLang/julia/pull/58948#issuecomment-3074321607) and corresponding [report](https://s3.amazonaws.com/julialang-reports/nanosoldier/pkgeval/by_hash/f175dd0_vs_228edd6/report.html). I have no idea how complicated it is to set something like this up, and I guess Github doesn't do it for free. Or maybe it is possible to include this kind of benchmark in CI? With up to 1M households, it currently only takes about 4 minutes to run the benchmark twice (for main and PR).