@hmgaudecker 2025-08-13
> (@hmgaudecker Regarding making a PR for the benchmark code in https://github.com/ttsim-dev/gettsim-code-for-picking: Is it okay if I push it with --no-verify? The code is 99% AI-generated and ruff almost gets a heart attack because of the ~100 "print"-calls...

Sure, though the pre-commit bot may complain... But nobody cares over there, once we move to routine performance checks as you detail above (maybe make an issue out of it? we should definitely do that at some point), we can still change that.

Thanks a lot!